requirements.txt
tavily-python>=0.3.2
PyYAML>=6.0.1
tenacity>=8.2.3
openai>=1.0.0
